Move items.py into components
