+++ ave/components/__init__.py
@@ -1,3 +1,5 @@
+"""Components of a parsed AVE game."""
+
 from .with_requirements import TextWithRequirements  # noqa: F401
 from .with_requirements import NameWithRequirements  # noqa: F401
 from .with_requirements import OptionWithRequirements  # noqa: F401
